# Labor.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Labor.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Order": {
      "post": {
        "tags": [
          "Order"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateOrderDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OrderCreationResponseDtoApiResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "orderStatus",
            "type": "string"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 20
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ObjectIEnumerableApiResponse"
            }
          }
        }
      }
    },
    "/api/Order/{orderId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OrderResponseDtoApiResponse"
            }
          }
        }
      }
    },
    "/api/Order/{orderId}/labor-summary": {
      "get": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/OrderLaborSummaryDtoApiResponse"
            }
          }
        }
      }
    },
    "/api/Order/{orderId}/status": {
      "put": {
        "tags": [
          "Order"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateOrderStatusDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/api/Order/items/{orderItemId}/status": {
      "put": {
        "tags": [
          "Order"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderItemId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateOrderItemStatusDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/api/Order/items/{orderItemId}/confirm": {
      "post": {
        "tags": [
          "Order"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderItemId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LaborConfirmAvailabilityDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/api/Order/{orderId}/tracking": {
      "post": {
        "tags": [
          "Order"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddOrderTrackingDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AddOrderTrackingDto": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "AddressDto": {
      "required": [
        "city",
        "country",
        "state",
        "street",
        "zipCode"
      ],
      "type": "object",
      "properties": {
        "addressId": {
          "format": "int32",
          "type": "integer"
        },
        "addressType": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "street": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "city": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "state": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "country": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "zipCode": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        },
        "isDefault": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "CreateOrderDto": {
      "required": [
        "workAddressId"
      ],
      "type": "object",
      "properties": {
        "workAddressId": {
          "format": "int32",
          "type": "integer"
        },
        "scheduledDate": {
          "format": "date-time",
          "type": "string"
        },
        "specialInstructions": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "LaborConfirmAvailabilityDto": {
      "type": "object",
      "properties": {
        "accepted": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "ObjectIEnumerableApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": { }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "OrderCreationResponseDto": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "orderNumber": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OrderCreationResponseDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/OrderCreationResponseDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "OrderItemDto": {
      "type": "object",
      "properties": {
        "orderItemId": {
          "format": "int32",
          "type": "integer"
        },
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "requiredHours": {
          "format": "int32",
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "workDescription": {
          "type": "string"
        },
        "itemStatus": {
          "type": "string"
        },
        "actualHours": {
          "format": "int32",
          "type": "integer"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        },
        "endTime": {
          "format": "date-time",
          "type": "string"
        },
        "laborName": {
          "type": "string"
        },
        "laborMobile": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "laborType": {
          "type": "string"
        },
        "specialization": {
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "OrderLaborItemDto": {
      "type": "object",
      "properties": {
        "orderItemId": {
          "format": "int32",
          "type": "integer"
        },
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "laborName": {
          "type": "string"
        },
        "laborMobile": {
          "type": "string"
        },
        "laborType": {
          "type": "string"
        },
        "requiredHours": {
          "format": "int32",
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "workDescription": {
          "type": "string"
        },
        "itemStatus": {
          "type": "string"
        },
        "preferredWorkDate": {
          "format": "date-time",
          "type": "string"
        },
        "confirmationStatus": {
          "type": "string"
        },
        "attemptCount": {
          "format": "int32",
          "type": "integer"
        },
        "respondedAt": {
          "format": "date-time",
          "type": "string"
        },
        "hasReview": {
          "type": "boolean"
        },
        "reviewId": {
          "format": "int32",
          "type": "integer"
        },
        "reviewRating": {
          "format": "int32",
          "type": "integer"
        },
        "reviewComment": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OrderLaborSummaryDto": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/OrderSummaryHeaderDto"
        },
        "laborItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderLaborItemDto"
          }
        }
      },
      "additionalProperties": false
    },
    "OrderLaborSummaryDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/OrderLaborSummaryDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "OrderResponseDto": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "orderNumber": {
          "type": "string"
        },
        "employerId": {
          "format": "int32",
          "type": "integer"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "orderStatus": {
          "type": "string"
        },
        "paymentStatus": {
          "type": "string"
        },
        "scheduledDate": {
          "format": "date-time",
          "type": "string"
        },
        "completedDate": {
          "format": "date-time",
          "type": "string"
        },
        "cancelledDate": {
          "format": "date-time",
          "type": "string"
        },
        "cancelReason": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "employerName": {
          "type": "string"
        },
        "employerMobile": {
          "type": "string"
        },
        "workAddress": {
          "$ref": "#/definitions/AddressDto"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItemDto"
          }
        },
        "orderTrackings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderTrackingDto"
          }
        }
      },
      "additionalProperties": false
    },
    "OrderResponseDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/OrderResponseDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "OrderSummaryHeaderDto": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "orderNumber": {
          "type": "string"
        },
        "orderStatus": {
          "type": "string"
        },
        "paymentStatus": {
          "type": "string"
        },
        "scheduledDate": {
          "format": "date-time",
          "type": "string"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "street": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        },
        "totalLabor": {
          "format": "int32",
          "type": "integer"
        },
        "confirmedLabor": {
          "format": "int32",
          "type": "integer"
        },
        "pendingLabor": {
          "format": "int32",
          "type": "integer"
        },
        "declinedLabor": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "OrderTrackingDto": {
      "type": "object",
      "properties": {
        "trackingId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateOrderItemStatusDto": {
      "type": "object",
      "properties": {
        "itemStatus": {
          "type": "string"
        },
        "actualHours": {
          "format": "int32",
          "type": "integer"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        },
        "endTime": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UpdateOrderStatusDto": {
      "type": "object",
      "properties": {
        "newStatus": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}